+++ src/result.rs
@@ -1,29 +1,40 @@
-use std::borrow::Cow;
+use std::fmt::Display;
 
-#[derive(thiserror::Error, Debug)]
+use miette::miette;
+
+#[derive(Debug)]
 pub enum Error {
-    #[error("The stream is not available")]
     UnavailableStream,
 
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
-
-    #[error(transparent)]
-    SerdeJson(#[from] serde_json::Error),
-
-    #[error("{0}")]
-    Msg(Cow<'static, str>),
-
-    #[error(transparent)]
-    Anyhow(#[from] anyhow::Error),
+    Miette(miette::Report),
 }
 
-pub fn err_msg(msg: impl Into<Cow<'static, str>>) -> Error {
-    Error::Msg(msg.into())
+impl From<miette::Report> for Error {
+    fn from(err: miette::Report) -> Self {
+        Error::Miette(err)
+    }
 }
 
-pub fn bail<T>(msg: impl Into<Cow<'static, str>>) -> Result<T> {
-    Err(err_msg(msg))
+impl From<Error> for miette::Report {
+    fn from(err: Error) -> Self {
+        match err {
+            Error::UnavailableStream => miette!("Unavailable stream"),
+            Error::Miette(err) => err,
+        }
+    }
+}
+
+impl Error {
+    pub fn wrap_err_with<D, F>(self, f: F) -> Error
+    where
+        D: Display + Send + Sync + 'static,
+        F: FnOnce() -> D,
+    {
+        match self {
+            Error::Miette(report) => Error::Miette(report.wrap_err(f())),
+            err => err,
+        }
+    }
 }
 
 pub type Result<T> = std::result::Result<T, Error>;